+++ code/TDC/TDC/IRepository/IAccountRepository.cs
@@ -1,0 +1,13 @@
+﻿using TDC.Models;
+
+namespace TDC.IRepository;
+public interface IAccountRepository
+{
+    public Account GetAccountById(long id);
+    public void CreateAccount(Account account);
+    public void UpdateAccount(Account account);
+    public void DeleteAccount(Account account);
+    public bool EmailIsTaken(string email);
+    public bool UsernameIsTaken(string username);
+
+}
