--- code/TDC/TDC/Repositories/ProfileRepository.cs
@@ -1,8 +1,0 @@
-﻿using TDC.Models;
-
-namespace TDC.Repositories;
-
-public class ProfileRepository
-{
-}
-
