+++ code/TDC/TDC/Repository/AccountRepository.cs
@@ -1,0 +1,76 @@
+﻿using TDC.IRepository;
+using TDC.Models;
+
+namespace TDC.Repositories;
+public class AccountRepository : IAccountRepository
+{
+    private readonly string projectPath = Path.GetFullPath(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, @"..\..\..\..\.."));
+    private readonly string filePath;
+
+    #region constructors
+    public AccountRepository()
+    {
+        filePath = Path.Combine(projectPath, "Data/Accounts");
+
+        #if ANDROID
+            var directoryPath = Environment.GetFolderPath(Environment.SpecialFolder.Personal);
+            if (!Directory.Exists(directoryPath))
+            {
+                Directory.CreateDirectory(directoryPath);
+            }
+
+            filePath = Path.Combine(directoryPath, "Accounts");
+        #endif
+    }
+    #endregion
+
+    #region publics
+    public void CreateAccount(Account account)
+    {
+        throw new NotImplementedException();
+    }
+
+    public void DeleteAccount(Account account)
+    {
+        throw new NotImplementedException();
+    }
+
+    public bool EmailIsTaken(string email)
+    {
+        throw new NotImplementedException();
+    }
+
+    public Account GetAccountById(long id)
+    {
+        throw new NotImplementedException();
+    }
+
+    public void UpdateAccount(Account account)
+    {
+        throw new NotImplementedException();
+    }
+
+    public bool UsernameIsTaken(string username)
+    {
+        throw new NotImplementedException();
+    }
+    #endregion
+
+    #region privates
+
+    // TO-DO: remove once database is implemented, only for android testing
+    private static List<Account> GetDummyAccounts()
+    {
+        var dummyAccounts = new List<Account>();
+        var acc1 = new Account(1, "acc1", "description1", "mail 1", "pw1", new Character("", "", 0, new CharacterStats(0, 0, 0)), [], []);
+        dummyAccounts.Add(acc1);
+
+        var acc2 = new Account(2, "acc2", "description2", "mail 2", "pw2", new Character("", "", 0, new CharacterStats(0, 0, 0)), [1], []);
+        dummyAccounts.Add(acc2);
+
+        var acc3 = new Account(3, "acc3", "description3", "mail 3", "pw3", new Character("", "", 0, new CharacterStats(0, 0, 0)), [2], [1]);
+        dummyAccounts.Add(acc3);
+        return dummyAccounts;
+    }
+    #endregion
+}
